requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-generativeai
yt-dlp
python-dotenv
apscheduler